requirements.txt
flet~=0.27.6
pandas~=2.2.3
email_validator~=2.2.0
beautifulsoup4~=4.13.3
python-dotenv~=1.1.0
IMAPClient~=3.0.1